Added support for building against rust beta channel.
"Dockerfile.beta" is currently temporary to build using musl builder. This will change in the future to have everything done without the need of an additional dockerfile. Also, while the beta channel is usually stable, it is best to keep in mind that there may be bugs in the beta channel that is not in the stable channel and may or may not affect the stability of the application.

closes #89