Change config to use sections instead.
With this change, there will be two sections called `[general]` and `[web-server]`. The general section will be required if a torchbear.toml file is going to be used, but if there is no file or if there is a file but the `web-server` section is missing, the server portion of torchbear will be disabled.