Updated install.sh
This change to the installer is mainly to improve the installation of torchbear on windows as well as make it a requirement to use cmder for the time being by checking for the variable `CMDER_ROOT`, which is exported at cmder runtime. Also changed how we check for the existance of torchbear as well.