Added `torchbear.os`
Using `torchbear.os` will return the os that torchbear is running on. If the system is not provided (eg if torchbear is ran on *bsd while there is no check for it) or is not available (eg running on a platform that dont have a valid os in `target_os`), it will return `unknown`.