Add dirs binding
This will be useful to access different common parts of the system on linux, windows, and mac (eg document, home, config, etc). There is a fallback builtin to the dirs crate that is used in the event that the variable used for the path is not found, and if nothing else works, it will return a nil on our end.