Implemented a common functions for io
`LuaCommonIO` will be used to share the same functions across multiple bindings dealing with io. This will reduce repeated code and allow for better functionality down the road. We are also making use of mutex allow safe usage, but in the future, we may change this to use a rwlock instead but shouldnt affect the overall functionality

Note: While we do have stderr, we currently arent reading from that at the moment. Also, while the functions have been tested to an extent and are working, they may change down the road.

TODO:
* Implement functions to read from stderr.
* Implement BufReader for stdout and stderr to add additional functions (eg read_line)