Update io bindings
Note: There is a regression between rlua and rust that cause degradation with some bindings (in this case, with io calls). Possibly related to how tantivy performance is affected. This was tested in both rlua 0.15 and 0.16, before and after the io changes, as well as external test. For now, we are to assume that many of the bindings are affected and will have a degrade in performance.

TODO: Further investigate into rlua and determine the best course to take to improve overall performance when creating a function with rlua that calls any rust functions.

Note related to #113