Implemented nanoid binding
NOTE: It seems there may be a breaking change in the crate that will use macros instead of the current functions. Unsure if they will bump minor version of the crate or not when they do release but its just something to keep in mind in case building torchbear fails in the future.

Close #253